+++ lib/ooxml_parser/docx_parser/document_structure/docx_paragraph/sdt/sdt_properties.rb
@@ -1,5 +1,10 @@
 # frozen_string_literal: true
 
+require_relative 'sdt_properties/form_properties'
+require_relative 'sdt_properties/form_text_properties'
+require_relative 'sdt_properties/combobox'
+require_relative 'sdt_properties/dropdown_list'
+require_relative 'sdt_properties/checkbox'
 module OoxmlParser
   # Class for parsing `w:sdtPr` tags
   class SDTProperties < OOXMLDocumentObject
@@ -9,6 +14,16 @@ module OoxmlParser
     attr_reader :tag
     # @return [ValuedChild] Locking Setting
     attr_reader :lock
+    # @return [ComboBox] combobox
+    attr_reader :combobox
+    # @return [DropdownList] dropdown list
+    attr_reader :dropdown_list
+    # @return [CheckBox] checkbox
+    attr_reader :checkbox
+    # @return [FormProperties] form properties
+    attr_reader :form_properties
+    # @return [FormTextProperties] properties of text in form
+    attr_reader :form_text_properties
 
     # Parse SDTProperties object
     # @param node [Nokogiri::XML:Element] node to parse
@@ -22,6 +37,16 @@ module OoxmlParser
           @tag = ValuedChild.new(:string, parent: self).parse(node_child)
         when 'lock'
           @lock = ValuedChild.new(:symbol, parent: self).parse(node_child)
+        when 'comboBox'
+          @combobox = ComboBox.new(parent: self).parse(node_child)
+        when 'dropDownList'
+          @dropdown_list = DropdownList.new(parent: self).parse(node_child)
+        when 'checkbox'
+          @checkbox = CheckBox.new(parent: self).parse(node_child)
+        when 'formPr'
+          @form_properties = FormProperties.new(parent: self).parse(node_child)
+        when 'textFormPr'
+          @form_text_properties = FormTextProperties.new(parent: self).parse(node_child)
         end
       end
       self